[IMP] account: improve Aged Partner reports performance
When generating e.g. Aged Receivable Report, _get_partner_move_lines
executed a first query, then a second query including a clause
with "id IN" potentially many records.
In local, for 75000 records, the function would take > 100 seconds.
For less than 100 records, it would take > 30 seconds.
We change _get_partner_move_lines so that the first query and the clause
are now in a subquery of the second query.
After the change, it always takes around 30 seconds.

opw 2119833

closes odoo/odoo#41385

Signed-off-by: Christophe Simonis <chs@odoo.com>